BUGGED - autocomplete update
The intended effect is for the user to type the name of their chosen Pokémon, and an autofill will be populated based on their entry. If they choose one of the autocomplete options, the user's input will change to that option and the query will be sent with that option.

What's happening right now is that the intended effect happens if the user clicks the autocomplete option a second time. The first click will send the user's partial input to be queried. A state update isn't working as intended. Will research and fix